Check chromosomes to plot

Check chromosomes to plot

Usage

.check_chroms(toplot, chrom_length, chroms, remove_empty_chrom)

Arguments

toplot

Tibble, main data-frame for plot.

chrom_length

Tibble of chromosome length.

chroms

Character vector, name of chromosomes to include in the plot. If NULL (default value), all chromosomes will be included.

remove_empty_chrom

Logical, should chromosomes with no significant markers/genes nor candidate genes be removed from the plot? Default value if FALSE.

Value

Character vector of chromosomes to plot.